integration of settings and map utility function
