+++ Abreath/.idea/misc.xml
@@ -4,24 +4,23 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="../../../../../layout/custom_preview.xml" value="0.3255208333333333" />
-        <entry key="..\:/Users/mpour/Documents/GitHub/breathalyzer/Abreath/app/src/main/res/layout/fragment_home.xml" value="0.18802083333333333" />
-        <entry key="..\:/Users/mpour/Documents/GitHub/breathalyzer/Abreath/app/src/main/res/layout/fragment_settings.xml" value="0.22" />
-        <entry key="..\:/Users/mpour/Documents/GitHub/breathalyzer/Abreath/app/src/main/res/layout/row.xml" value="0.36666666666666664" />
-        <entry key="..\:/Users/mpour/Documents/GitHub/breathalyzer/Abreath/app/src/main/res/menu/bottom_nav_menu.xml" value="0.18802083333333333" />
         <entry key="app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.303" />
         <entry key="app/src/main/res/drawable/gradient_header.xml" value="0.216" />
         <entry key="app/src/main/res/drawable/ic_baseline_settings_24.xml" value="0.384" />
         <entry key="app/src/main/res/drawable/ic_dashboard_black_24dp.xml" value="0.303" />
         <entry key="app/src/main/res/drawable/ic_home_black_24dp.xml" value="0.29" />
         <entry key="app/src/main/res/drawable/ic_launcher_background.xml" value="0.303" />
+        <entry key="app/src/main/res/drawable/overlay.xml" value="0.302" />
         <entry key="app/src/main/res/drawable/round_container.xml" value="0.352" />
         <entry key="app/src/main/res/layout-v31/fragment_home.xml" value="0.35052083333333334" />
+        <entry key="app/src/main/res/layout/activity_about_page.xml" value="0.20018115942028986" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.22" />
         <entry key="app/src/main/res/layout/fragment_dashboard.xml" value="0.33" />
         <entry key="app/src/main/res/layout/fragment_home.xml" value="0.33" />
         <entry key="app/src/main/res/layout/fragment_notifications.xml" value="0.5359375" />
         <entry key="app/src/main/res/layout/fragment_profile_graph.xml" value="0.1" />
         <entry key="app/src/main/res/layout/fragment_settings.xml" value="0.2994791666666667" />
+        <entry key="app/src/main/res/layout/row.xml" value="0.20018115942028986" />
         <entry key="app/src/main/res/menu/bottom_nav_menu.xml" value="0.2994791666666667" />
       </map>
     </option>
