+++ Abreath/app/src/main/java/com/coen390/abreath/ui/model/SettingsViewModel.java
@@ -1,16 +1,16 @@
-package com.coen390.abreath.ui.dashboard;
+package com.coen390.abreath.ui.model;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class DashboardViewModel extends ViewModel {
+public class SettingsViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public DashboardViewModel() {
+    public SettingsViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is dashboard fragment");
+        mText.setValue("This is the settings fragment");
     }
 
     public LiveData<String> getText() {
