+++ Abreath/app/src/main/java/com/coen390/abreath/ui/settings/SettingsAdapter.java
@@ -1,4 +1,4 @@
-package com.coen390.abreath;
+package com.coen390.abreath.ui.settings;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -10,6 +10,8 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+
+import com.coen390.abreath.R;
 
 import java.util.ArrayList;
