+++ Abreath/app/src/main/res/drawable/overlay.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="@color/primaryColor" android:dither="true"/>
+</shape>
