+++ Abreath/app/src/main/res/layout/fragment_settings.xml
@@ -8,26 +8,13 @@
 
     <ListView
         android:id="@+id/listView"
-        android:layout_width="328dp"
+        android:layout_width="match_parent"
+        android:layout_marginStart="17dp"
+        android:layout_marginEnd="17dp"
         android:layout_height="438dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <TextView
-        android:id="@+id/settings"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="180dp"
-        android:layout_marginTop="16dp"
-        android:layout_marginEnd="180dp"
-        android:text="Settings"
-        android:textColor="#000000"
-        android:textColorLink="#000000"
-        android:textSize="24sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        android:textStyle="bold"/>
 </androidx.constraintlayout.widget.ConstraintLayout>
