+++ Abreath/app/src/main/res/layout/row.xml
@@ -23,14 +23,16 @@
         android:layout_marginTop="5dp"
         android:layout_toRightOf="@+id/image"
         android:layout_marginLeft="5dp"
-        android:textColor="@color/black"/>
+        />
 
     <ImageView
         android:id="@+id/arrow"
         android:layout_width="35dp"
         android:layout_height="35dp"
-        android:layout_marginHorizontal="290dp"
 
+
+        android:layout_alignParentEnd="true"
+        android:layout_marginEnd="8dp"
         android:src="@drawable/arrow" />
 
 </RelativeLayout>
